chore: include hotfix-15958
