+++ dotCMS/src/main/java/com/dotmarketing/portlets/browser/ajax/BrowserAjax.java
@@ -371,148 +371,9 @@ public class BrowserAjax {
 		Map<String, Object> results = browserAPI.getFolderContent(usr,
 				folderId, offset, maxResults, filter, mimeTypes, extensions,
 				showArchived, noFolders, onlyFiles, sortBy, sortByDesc,
-				excludeLinks, getAllLanguages);
-
-		listCleanup((List<Map<String, Object>>) results.get("list"), getContentSelectedLanguageId(req));
+				excludeLinks, getAllLanguages, getContentSelectedLanguageId(req));
 
 		return results;
-	}
-
-	/**
-	 * The list of content under a folder might include the identifier
-	 * several times, representing all the available languages for a single
-	 * content.
-	 * <p>
-	 * This method takes that list and <i>leaves only one identifier per
-	 * page</i>. This unique record represents either the content with the default
-	 * language ID, or the content with the next language ID in the list of system
-	 * languages.
-	 * </p>
-     * * <p>
-     * If the fallback properties are false <i>leaves only one identifier per
-     * page that match the param languageId</i>.
-     * </p>
-	 * 
-	 * @param results
-	 *            - The full list of pages under a given path/directory.
-     * @param languageId
-     *            - Content Language of the results.
-	 */
-	private void listCleanup(List<Map<String, Object>> results, long languageId) {
-		Map<String, Integer> contentLangCounter = new HashMap<>();
-
-		// Examine only the pages with more than 1 assigned language
-		for (Map<String, Object> content : results) {
-			if ((boolean) content.get("isContentlet")) {
-				String ident = (String) content.get("identifier");
-				if (contentLangCounter.containsKey(ident)) {
-					int counter = contentLangCounter.get(ident);
-					contentLangCounter.put(ident, counter + 1);
-				} else {
-					contentLangCounter.put(ident, 1);
-				}
-			}
-		}
-
-		Set<String> identifierSet = contentLangCounter.keySet();
-		for (String identifier : identifierSet) {
-			int counter = contentLangCounter.get(identifier);
-			if (counter > 1) {
-				// Remove all languages except the default one
-				boolean isDeleted = removeAdditionalLanguages(identifier, results, languageId);
-				if (!isDeleted) {
-					// Otherwise, leave only the next available language
-					List<Language> languages = this.languageAPI.getLanguages();
-					for (Language language : languages) {
-						if (language.getId() != languageId) {
-							isDeleted = removeAdditionalLanguages(identifier, results, language.getId());
-							if (isDeleted) {
-								break;
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-
-	/**
-	 * Removes all other contents from the given list that ARE NOT associated to
-	 * the specified language ID. In the end, the list will contain one content per
-	 * identifier with either the default language ID or the next available
-	 * language.
-	 * 
-	 * @param identifier
-	 *            - The identifier of the page to clean up in the list.
-	 * @param resultList
-	 *            - The list of all pages that will be displayed.
-	 * @param languageId
-	 *            - The language ID of the page that will remain in the list.
-	 * @return If <code>true</code>, the identifier with the specified language
-	 *         ID was successfully cleaned up. If <code>false</code>, the
-	 *         identifier is not associated to the specified language ID and was
-	 *         not removed from the list.
-	 */
-	private boolean removeAdditionalLanguages(String identifier,
-			List<Map<String, Object>> resultList, long languageId) {
-
-		boolean removeOtherLangs = false;
-
-		for (Map<String, Object> contentInfo : resultList) {
-			if ((boolean) contentInfo.get("isContentlet")) {
-				String ident = (String) contentInfo.get("identifier");
-				if (identifier.equals(ident)) {
-					long langId = (long) contentInfo.get("languageId");
-					// If specified language is found, remove all others
-					if (languageId == langId) {
-						removeOtherLangs = true;
-						break;
-					}
-				}
-			}
-		}
-
-		if (removeOtherLangs) {
-			removeLangOtherThan(resultList, identifier, languageId);
-		}
-
-		return removeOtherLangs;
-	}
-
-	/**
-	 * Removes all the pages from the list that are not associated to the
-	 * specified language. In the end, the list will contain one page per
-	 * identifier.
-	 * 
-	 * @param resultList
-	 *            - The list of pages that will be displayed.
-	 * @param identifier
-	 *            - The identifier for the page to lookup in the list.
-	 * @param languageId
-	 *            - The language ID that <b>MUST REMAIN</b> in the list.
-	 */
-	private void removeLangOtherThan(List<Map<String, Object>> resultList,
-			String identifier, long languageId) {
-		List<Integer> itemsToRemove = new ArrayList<Integer>();
-		for (int i = 0; i < resultList.size(); i++) {
-			Map<String, Object> pageInfo = resultList.get(i);
-			if ((boolean) pageInfo.get("isContentlet")) {
-				String ident = (String) pageInfo.get("identifier");
-				if (identifier.equals(ident)) {
-					long langId = (long) pageInfo.get("languageId");
-					if (languageId != langId) {
-						itemsToRemove.add(i);
-					}
-				}
-			}
-		}
-		int deletionCounter = 0;
-		for (int index : itemsToRemove) {
-			// Adjust index based on previous deletions
-			int indexAfterDeletion = index - deletionCounter;
-			resultList.remove(indexAfterDeletion);
-			deletionCounter++;
-		}
 	}
 
 	public Map<String, Object> saveFileAction(String selectedItem,String wfActionAssign,String wfActionId,String wfActionComments, String wfConId, String wfPublishDate,
